requirements.txt
textual[dev]==0.5.0
pyyaml
rich
python-dotenv
pyte
